@@ -12,7 +12,7 @@
 	"license": "AGPL-3.0",
 	"repository": {
 		"type": "git",
-		"url": "https://github.com/pokegh0st/pokebook"
+		"url": "https://github.com/pokeghosst/pokebook"
 	},
 	"type": "module",
 	"scripts": {
@@ -30,7 +30,7 @@
 		"@capacitor/assets": "^3.0.5",
 		"@sveltejs/adapter-auto": "^3.2.3",
 		"@sveltejs/adapter-static": "^3.0.3",
-		"@sveltejs/kit": "^2.5.21",
+		"@sveltejs/kit": "^2.5.27",
 		"@tauri-apps/cli": "^1.6.0",
 		"@typescript-eslint/eslint-plugin": "^5.45.0",
 		"@typescript-eslint/parser": "^5.45.0",
@@ -38,14 +38,14 @@
 		"dotenv": "^16.4.5",
 		"eslint": "^8.54.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.43.0",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.8.1",
+		"eslint-plugin-svelte": "^2.45.1",
+		"prettier": "^3.1.0",
+		"prettier-plugin-svelte": "^3.2.6",
 		"sass": "^1.77.8",
-		"svelte": "^4.2.18",
-		"typescript": "^4.9.5",
+		"svelte": "^5.19.1",
+		"typescript": "^5.5.0",
 		"typescript-eslint": "^8.1.0",
-		"vite": "^5.4.0",
+		"vite": "^6.0.11",
 		"vite-plugin-top-level-await": "^1.4.4"
 	},
 	"dependencies": {
