@@ -67,8 +67,8 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 	</ul>
 	<div class="about-bottom">
 		<p>
-			<a href="/privacy" on:click={closeModal}>{$t('about.privacy')}</a>
-			| <a href="/terms" on:click={closeModal}>{$t('about.terms')}</a>
+			<a href="/privacy" onclick={closeModal}>{$t('about.privacy')}</a>
+			| <a href="/terms" onclick={closeModal}>{$t('about.terms')}</a>
 		</p>
 		<p>© {new Date().getFullYear()} Pokeghost</p>
 	</div>
