@@ -17,6 +17,9 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 -->
 
 <script lang="ts">
+	import { run, createBubbler } from 'svelte/legacy';
+
+	const bubble = createBubbler();
 	import { openModal } from 'svelte-modals';
 
 	import { isSidebarOpen } from '$lib/stores/isSidebarOpen';
@@ -35,27 +38,29 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 		}
 	}
 
-	let sidebarNavOpenClass = '';
+	let sidebarNavOpenClass = $state('');
 
-	$: $isSidebarOpen === 'true'
-		? (sidebarNavOpenClass = 'sidebar-nav--open')
-		: (sidebarNavOpenClass = '');
+	run(() => {
+		$isSidebarOpen === 'true'
+			? (sidebarNavOpenClass = 'sidebar-nav--open')
+			: (sidebarNavOpenClass = '');
+	});
 </script>
 
 <div class="sidebar-nav-wrapper">
 	<div
 		class="sidebar-close-area {sidebarNavOpenClass}"
-		on:click={() => ($isSidebarOpen = 'false')}
-		on:keydown
+		onclick={() => ($isSidebarOpen = 'false')}
+		onkeydown={bubble('keydown')}
 		role="button"
 		tabindex="0"
-	/>
+	></div>
 	<div class="sidebar {sidebarNavOpenClass}">
 		<div class="sidebar-nav-items">
 			{#each navMenuItems as item, index (navMenuItems[index])}
-				<a href={item.url} on:click={() => handleSidebarItemClick()}>
+				<a href={item.url} onclick={() => handleSidebarItemClick()}>
 					<div class="list-item">
-						<svelte:component this={item.icon} />
+						<item.icon />
 						{$t(item.label)}
 					</div>
 				</a>
@@ -63,12 +68,12 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 		</div>
 		<div class="sidebar-footer">
 			<button
-				on:click={() => openModal(Modal, { title: $t('workspace.hotkeys'), content: HotkeysModal })}
+				onclick={() => openModal(Modal, { title: $t('workspace.hotkeys'), content: HotkeysModal })}
 				>{$t('menu.shortcuts')}</button
 			>
 			<ul>
 				<li>
-					<button on:click={() => openModal(Modal, { content: AboutModal })}
+					<button onclick={() => openModal(Modal, { content: AboutModal })}
 						>{$t('menu.about')}</button
 					>
 				</li>
