@@ -36,8 +36,8 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 	const FALLBACK_DELAY_MS = 100;
 
-	let cachedPoems: Promise<PoemCacheRecord[]>;
-	let showFallback = false;
+	let cachedPoems: Promise<PoemCacheRecord[]> = $state();
+	let showFallback = $state(false);
 	let fallbackTimeout: ReturnType<typeof setTimeout>;
 
 	onMount(() => {
@@ -69,12 +69,12 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 {:then cacheRecords}
 	{#if cacheRecords && cacheRecords.length > 0}
 		<div class="refresh-wrapper">
-			<button class="button" on:click={handleCacheRefresh}>Refresh <RotateCcw /></button>
+			<button class="button" onclick={handleCacheRefresh}>Refresh <RotateCcw /></button>
 		</div>
 		<div class="poem-list">
 			{#each cacheRecords as record}
 				<div class="list-item">
-					<button on:click={() => goToPoem(record.id)}>
+					<button onclick={() => goToPoem(record.id)}>
 						<div class="list-poem">
 							<p class="list-poem-name">
 								{record.name}{record.unsavedChanges ? ` (${$t('workspace.unsaved')})` : ''}
